# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/jobs/{jobId}/documents": {
      "post": {
        "operationId": "DocumentsController_generate",
        "parameters": [
          {
            "name": "jobId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "documents"
        ]
      }
    },
    "/jobs/{jobId}/documents/{docId}": {
      "get": {
        "operationId": "DocumentsController_getOne",
        "parameters": [
          {
            "name": "jobId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "docId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "documents"
        ]
      }
    },
...
  },
  "info": {
    "title": "WinProp API",
    "description": "",
    "version": "1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "DocumentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "jobId": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "proposal"
            ]
          },
          "title": {
            "type": "string"
          },
          "contentJson": {
            "type": "object",
            "additionalProperties": true
          },
          "version": {
            "type": "number"
          },
          "status": {
            "enum": [
              "draft",
              "ready"
            ],
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "jobId",
          "type",
          "title",
          "contentJson",
          "version",
          "status",
          "createdAt"
        ]
      },
...
    }
  }
}